refactor: split custom data table from modal #1097
