+++ libs/frontend/feature-custom-data/tsconfig.json
@@ -6,7 +6,7 @@
     "declaration": true,
     "target": "ESNext"
   },
-  "include": ["**/*.ts"],
+  "include": ["**/*.ts", "**/*.vue"],
   "exclude": ["**/*.spec.ts"],
   "references": [
     {
