+++ libs/frontend/ui-widgets/src/lib/TableNav.vue
